# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"🇻🇳 Vietnamese Multi-Document Summarization API","description":"Hệ thống tóm tắt văn bản tiếng Việt đa tài liệu.\n\nHỗ trợ:\n- **Extractive**: Trích xuất câu quan trọng (TextRank)\n- **Abstractive**: Sinh câu mới (ViT5 Transformer)\n- **ROUGE Evaluation**: Đánh giá và chọn bản tóm tắt tốt nhất","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Info"],"summary":"Root","description":"Thông tin API và danh sách endpoints.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Info"],"summary":"Health Check","description":"Kiểm tra trạng thái server và model.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/summarize":{"post":{"tags":["Summarization"],"summary":"Tóm tắt văn bản tiếng Việt","description":"**Tóm tắt văn bản tiếng Việt** từ text trực tiếp hoặc danh sách URLs.\n\nQuy trình xử lý:\n1. Crawl URLs (nếu có) và gộp với text trực tiếp\n2. Tiền xử lý: làm sạch HTML, chuẩn hóa Unicode\n3. Tóm tắt trích xuất (TextRank)\n4. Tóm tắt diễn giải (ViT5)\n5. Đánh giá ROUGE và chọn bản tốt nhất\n\n**Lưu ý**: Cần ít nhất một trong hai: `text` hoặc `urls`.","operationId":"summarize_summarize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/summarize/files":{"post":{"tags":["Summarization"],"summary":"Upload TXT/PDF/DOCX và tóm tắt nhiều tài liệu","operationId":"summarize_files_summarize_files_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_summarize_files_summarize_files_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_summarize_files_summarize_files_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference"},"extractive_sentences":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Extractive Sentences","default":5},"max_abstractive_length":{"type":"integer","maximum":512.0,"minimum":30.0,"title":"Max Abstractive Length","default":150},"length_control":{"type":"string","title":"Length Control","default":"auto"},"model_name":{"type":"string","title":"Model Name","default":"vit5"},"analysis_mode":{"type":"string","title":"Analysis Mode","default":"fast"},"save_result":{"type":"boolean","title":"Save Result","default":true}},"type":"object","required":["files"],"title":"Body_summarize_files_summarize_files_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SummarizeRequest":{"properties":{"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text","description":"Văn bản tiếng Việt trực tiếp cần tóm tắt","example":"Hội đồng Bảo an Liên Hợp Quốc đã họp khẩn cấp..."},"urls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Urls","description":"Danh sách URL bài báo cần crawl và tóm tắt","example":["https://vnexpress.net/some-article"]},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference","description":"Văn bản tham chiếu để tính ROUGE (tuỳ chọn). Nếu không cung cấp, sẽ dùng văn bản gốc làm reference."},"extractive_sentences":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Extractive Sentences","description":"Số câu trong bản tóm tắt trích xuất","default":5},"max_abstractive_length":{"type":"integer","maximum":512.0,"minimum":30.0,"title":"Max Abstractive Length","description":"Độ dài tối đa bản tóm tắt diễn giải (số token)","default":150},"length_control":{"type":"string","title":"Length Control","description":"auto | 20_percent | 50_percent | 100_words | 200_words","default":"auto"},"model_name":{"type":"string","title":"Model Name","description":"vit5/t5 hoặc bart. Có thể truyền tên model Hugging Face bất kỳ.","default":"vit5"},"save_result":{"type":"boolean","title":"Save Result","description":"Lưu kết quả ra storage JSON và MongoDB nếu MONGO_URI được cấu hình.","default":true},"analysis_mode":{"type":"string","title":"Analysis Mode","description":"fast hoặc full. Fast chỉ kiểm chứng top summary sentences để giảm thời gian.","default":"fast"}},"type":"object","title":"SummarizeRequest","description":"Schema cho request POST /summarize","example":{"extractive_sentences":3,"max_abstractive_length":100,"text":"Hội đồng Bảo an Liên Hợp Quốc đã họp khẩn cấp để thảo luận về tình hình leo thang căng thẳng ở Trung Đông. Nhiều quốc gia kêu gọi ngừng bắn ngay lập tức và mở hành lang nhân đạo cho người dân vùng chiến sự.","urls":[]}},"SummarizeResponse":{"properties":{"extractive":{"type":"string","title":"Extractive","description":"Bản tóm tắt trích xuất (TextRank)"},"abstractive":{"type":"string","title":"Abstractive","description":"Bản tóm tắt diễn giải (ViT5)"},"best":{"type":"string","title":"Best","description":"Bản tóm tắt được chọn (tốt nhất)"},"best_type":{"type":"string","title":"Best Type","description":"Loại tóm tắt tốt nhất: 'extractive' hoặc 'abstractive'"},"scores":{"additionalProperties":true,"type":"object","title":"Scores","description":"Điểm ROUGE chi tiết cho cả hai bản tóm tắt"},"word_count":{"additionalProperties":true,"type":"object","title":"Word Count","description":"Số từ của mỗi bản tóm tắt"},"processing_time_seconds":{"type":"number","title":"Processing Time Seconds","description":"Thời gian xử lý (giây)"},"consistency":{"additionalProperties":true,"type":"object","title":"Consistency"},"explainability":{"additionalProperties":true,"type":"object","title":"Explainability"},"documents":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Documents"},"storage":{"additionalProperties":true,"type":"object","title":"Storage"},"controls":{"additionalProperties":true,"type":"object","title":"Controls"}},"type":"object","required":["extractive","abstractive","best","best_type","scores","word_count","processing_time_seconds"],"title":"SummarizeResponse","description":"Schema cho response POST /summarize"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}